% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapro.R
\name{am.control}
\alias{am.control}
\title{Add Control}
\usage{
am.control(id, ctype = NULL, ...)
}
\arguments{
\item{id}{\code{amapro} id or widget from \link{am.init}}

\item{ctype}{A string for name of control, like 'Scale','ControlBar','ToolBar'.}

\item{...}{A named list of parameters for the chosen control}
}
\value{
A map widget to plot, or to save and expand with more features.
}
\description{
Add a Control to a map.
}
\details{
controls are ControlBar, ToolBar and Scale. \cr
\href{https://a.amap.com/jsapi/static/doc/20210906/index.html?v=2#control}{Parameters} could be position or offset.\cr
}
\examples{
if (interactive()) {
  am.init() |> am.control("Scale")
}
}
\seealso{
\link{am.init} code example
}
