% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapro.R
\name{am.inspect}
\alias{am.inspect}
\title{Map to JSON}
\usage{
am.inspect(wt, json = TRUE, ...)
}
\arguments{
\item{wt}{An \code{amapro} widget as returned by \link{am.init}}

\item{json}{Boolean whether to return a JSON, or a \code{list}, default TRUE}

\item{...}{Additional arguments to pass to \link[jsonlite]{toJSON}}
}
\value{
A JSON string if \code{json} is \code{TRUE} and
a \code{list} otherwise.
}
\description{
Convert map elements to JSON string
}
\details{
Must be invoked or chained as last command.\cr
}
\examples{
am.init(viewMode= '3D', zoom= 10, pitch= 60) |>
  am.control(ctype= 'ControlBar', position= 'RT') |>
  am.inspect()

}
