% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapro.R
\name{am.item}
\alias{am.item}
\title{Add Item}
\usage{
am.item(id, itype, ...)
}
\arguments{
\item{id}{A valid widget from \link{am.init}}

\item{itype}{A string for item type name, like 'Marker'}

\item{...}{attributes of item}
}
\value{
A map widget to plot, or to save and expand with more features
}
\description{
Add an item to a map
}
\details{
To add an item like Marker, Text or Polyline to the map
}
\examples{
if (interactive()) {
  am.init() |> am.item('Marker', position=c(116.6, 40))
}
}
\seealso{
\link{am.init} code example
}
