% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapro.R
\name{am.output}
\alias{am.output}
\title{Shiny: map UI}
\usage{
am.output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Name of output UI element.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \emph{'auto'}) or a number, which will be coerced to a
string and have \emph{'px'} appended.}
}
\value{
An output or render function that enables the use of the widget within Shiny applications.
See \link[htmlwidgets]{shinyWidgetOutput}.
}
\description{
Placeholder for a map in Shiny UI
}
\seealso{
Shiny demo in demo(am.shiny)
}
