% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapro.R
\name{am.proxy}
\alias{am.proxy}
\title{Shiny: create a map proxy}
\usage{
am.proxy(id)
}
\arguments{
\item{id}{Target map id from the Shiny UI}
}
\value{
A proxy object to update the map
}
\description{
Create a proxy for an existing map in Shiny. It allows to
add, merge, delete elements to a map without reloading it.
}
\examples{
if (interactive()) {
  demo(am.shiny)
}
}
