% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amapro.R
\name{am.render}
\alias{am.render}
\title{Shiny: render a map}
\usage{
am.render(wt, env = parent.frame())
}
\arguments{
\item{wt}{An \code{amapro} widget to generate the chart.}

\item{env}{The environment in which to evaluate \code{expr}.}
}
\value{
An output or render function that enables the use of the widget within Shiny applications.
}
\description{
This is the initial rendering of a map in the UI.
}
\seealso{
\link{am.proxy} for example, \code{\link[htmlwidgets]{shinyRenderWidget}} for return value.
}
